# api/openapi.json
{"openapi": "3.1.0", "info": {"title": "JuliaAI Backend Application", "version": "1.0.0"}, "paths": {"/api/v1/user/me": {"get": {"summary": "Userinfo", "operationId": "userinfo_api_v1_user_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/MilitaryUserRead"}, {"$ref": "#/components/schemas/TherapistUserRead"}], "title": "Response Userinfo Api V1 User Me Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/user": {"put": {"summary": "Update User", "operationId": "update_user_api_v1_user_put", "requestBody": {"content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/MilitaryUserUpdate"}, {"$ref": "#/components/schemas/TherapistUserUpdate"}], "title": "Body"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/MilitaryUserRead"}, {"$ref": "#/components/schemas/TherapistUserRead"}], "title": "Response Update User Api V1 User Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"summary": "Create User", "operationId": "create_user_api_v1_user_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryUserCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryUserRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/user/{user_id}": {"delete": {"summary": "Delete User", "operationId": "delete_user_api_v1_user__user_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/user/therapist": {"post": {"summary": "Create Therapist", "operationId": "create_therapist_api_v1_user_therapist_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TherapistCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TherapistUserRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/user/profile": {"put": {"summary": "Update Profile", "operationId": "update_profile_api_v1_user_profile_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryProfileUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryProfileRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"summary": "Create Profile", "operationId": "create_profile_api_v1_user_profile_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryProfileCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryProfileRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/user/{user_id}/profile": {"get": {"summary": "Get Profile", "operationId": "get_profile_api_v1_user__user_id__profile_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MilitaryProfileRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/assessment": {"get": {"summary": "Get Entry Test", "operationId": "get_entry_test_api_v1_assessment_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TestRead"}}}}}}, "post": {"summary": "Submit Entry Test", "operationId": "submit_entry_test_api_v1_assessment_post", "requestBody": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TestQuestionResult"}, "type": "array", "title": "Result"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TherapistProfileView"}, "type": "array", "title": "Response Submit Entry Test Api V1 Assessment Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/assessment/results": {"get": {"summary": "Get Recommendations", "operationId": "get_recommendations_api_v1_assessment_results_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TherapistProfileView"}, "type": "array", "title": "Response Get Recommendations Api V1 Assessment Results Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/chats": {"get": {"summary": "List Chats", "operationId": "list_chats_api_v1_chats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ChatPreviewRead"}, "type": "array", "title": "Response List Chats Api V1 Chats Get"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"summary": "Create Chat", "operationId": "create_chat_api_v1_chats_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/chats/{chat_id}": {"get": {"summary": "Get Chat", "operationId": "get_chat_api_v1_chats__chat_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/chats/{chat_id}/messages": {"post": {"summary": "Create Message", "operationId": "create_message_api_v1_chats__chat_id__messages_post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_create_message_api_v1_chats__chat_id__messages_post"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessagePreviewRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"summary": "List Messages", "operationId": "list_messages_api_v1_chats__chat_id__messages_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/MessagePreviewRead"}, "title": "Response List Messages Api V1 Chats  Chat Id  Messages Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/chats/{chat_id}/messages/{message_id}": {"delete": {"summary": "Delete Message", "operationId": "delete_message_api_v1_chats__chat_id__messages__message_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}, {"name": "message_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Message Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_create_message_api_v1_chats__chat_id__messages_post": {"properties": {"message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["message"], "title": "Body_create_message_api_v1_chats__chat_id__messages_post"}, "ChatCreate": {"properties": {"participant_id": {"type": "string", "title": "Participant Id"}}, "type": "object", "required": ["participant_id"], "title": "ChatCreate"}, "ChatPreviewRead": {"properties": {"id": {"type": "string", "title": "Id"}, "created_at": {"type": "string", "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "participant": {"anyOf": [{"$ref": "#/components/schemas/MilitaryUserRead"}, {"$ref": "#/components/schemas/TherapistUserRead"}, {"type": "null"}], "title": "Participant"}, "last_updated_at": {"type": "string", "title": "Last Updated At"}, "last_message": {"anyOf": [{"$ref": "#/components/schemas/MessagePreviewRead"}, {"type": "null"}]}}, "type": "object", "required": ["id", "created_at", "updated_at", "participant", "last_updated_at", "last_message"], "title": "ChatPreviewRead"}, "ChatRead": {"properties": {"id": {"type": "string", "title": "Id"}, "created_at": {"type": "string", "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "participant": {"anyOf": [{"$ref": "#/components/schemas/MilitaryUserRead"}, {"$ref": "#/components/schemas/TherapistUserRead"}, {"type": "null"}], "title": "Participant"}, "last_updated_at": {"type": "string", "title": "Last Updated At"}, "last_message_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Message Id"}}, "type": "object", "required": ["id", "created_at", "updated_at", "participant", "last_updated_at", "last_message_id"], "title": "ChatRead"}, "Gender": {"type": "string", "enum": ["\u0427\u043e\u043b\u043e\u0432\u0456\u043a", "\u0416\u0456\u043d\u043a\u0430"], "title": "Gender"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MarriageStatus": {"type": "string", "enum": ["\u041d\u0435\u043e\u0434\u0440\u0443\u0436\u0435\u043d\u0438\u0439/\u043d\u0435\u0437\u0430\u043c\u0456\u0436\u043d\u044f", "\u041e\u0434\u0440\u0443\u0436\u0435\u043d\u0438\u0439/\u0437\u0430\u043c\u0456\u0436\u043d\u044f", "\u0420\u043e\u0437\u043b\u0443\u0447\u0435\u043d(\u0438\u0439/\u0430)", "\u0416\u0438\u0432\u0443 \u043e\u043a\u0440\u0435\u043c\u043e", "\u0412\u0434\u043e\u0432\u0430/\u0432\u0434\u0456\u0432\u0435\u0446\u044c"], "title": "MarriageStatus"}, "MessagePreviewRead": {"properties": {"id": {"type": "string", "title": "Id"}, "chat_id": {"type": "string", "title": "Chat Id"}, "author_id": {"type": "string", "title": "Author Id"}, "author_name": {"type": "string", "title": "Author Name"}, "author_role": {"anyOf": [{"$ref": "#/components/schemas/UserRole"}, {"type": "null"}]}, "content": {"type": "string", "title": "Content"}, "sent_at": {"type": "string", "title": "Sent At"}}, "type": "object", "required": ["id", "chat_id", "author_id", "author_name", "author_role", "content", "sent_at"], "title": "MessagePreviewRead"}, "MilitaryProfileCreate": {"properties": {"gender": {"anyOf": [{"$ref": "#/components/schemas/Gender"}, {"type": "null"}]}, "about": {"type": "string", "title": "About"}, "marriage_status": {"anyOf": [{"$ref": "#/components/schemas/MarriageStatus"}, {"type": "null"}]}, "kids": {"type": "integer", "title": "Kids", "default": 0}, "hobbies": {"type": "string", "title": "Hobbies", "default": ""}, "complains": {"type": "string", "title": "Complains"}, "insecures": {"type": "string", "title": "Insecures"}, "anxieties": {"type": "string", "title": "Anxieties"}}, "type": "object", "required": ["about", "complains", "insecures", "anxieties"], "title": "MilitaryProfileCreate"}, "MilitaryProfileRead": {"properties": {"created_at": {"type": "string", "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "user_id": {"type": "string", "title": "User Id"}, "gender": {"anyOf": [{"$ref": "#/components/schemas/Gender"}, {"type": "null"}]}, "about": {"type": "string", "title": "About"}, "marriage_status": {"anyOf": [{"$ref": "#/components/schemas/MarriageStatus"}, {"type": "null"}]}, "kids": {"type": "integer", "title": "Kids", "default": 0}, "hobbies": {"type": "string", "title": "Hobbies", "default": ""}, "complains": {"type": "string", "title": "Complains"}, "insecures": {"type": "string", "title": "Insecures"}, "anxieties": {"type": "string", "title": "Anxieties"}}, "type": "object", "required": ["created_at", "updated_at", "user_id", "gender", "about", "marriage_status", "complains", "insecures", "anxieties"], "title": "MilitaryProfileRead"}, "MilitaryProfileUpdate": {"properties": {"gender": {"anyOf": [{"$ref": "#/components/schemas/Gender"}, {"type": "null"}]}, "about": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "About"}, "marriage_status": {"anyOf": [{"$ref": "#/components/schemas/MarriageStatus"}, {"type": "null"}]}, "kids": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Kids"}, "hobbies": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Hobbies"}, "complains": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Complains"}, "insecures": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Insecures"}, "anxieties": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Anxieties"}}, "type": "object", "title": "MilitaryProfileUpdate"}, "MilitaryUserCreate": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "age": {"type": "integer", "title": "Age"}, "location": {"type": "string", "title": "Location"}, "rank": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Rank"}}, "type": "object", "required": ["first_name", "last_name", "age", "location"], "title": "MilitaryUserCreate"}, "MilitaryUserRead": {"properties": {"created_at": {"type": "string", "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "id": {"type": "string", "title": "Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "age": {"type": "integer", "title": "Age"}, "location": {"type": "string", "title": "Location"}, "role": {"$ref": "#/components/schemas/UserRole"}, "rank": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Rank"}}, "type": "object", "required": ["created_at", "updated_at", "id", "first_name", "last_name", "email", "phone", "age", "location", "role", "rank"], "title": "MilitaryUserRead"}, "MilitaryUserUpdate": {"properties": {"first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "age": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Age"}, "location": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Location"}, "rank": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Rank"}}, "type": "object", "title": "MilitaryUserUpdate"}, "TestQuestionRead": {"properties": {"id": {"type": "integer", "title": "Id"}, "content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["id", "content"], "title": "TestQuestionRead"}, "TestQuestionResult": {"properties": {"id": {"type": "integer", "title": "Id"}, "answer": {"type": "integer", "maximum": 5.0, "minimum": 1.0, "title": "Answer"}}, "type": "object", "required": ["id", "answer"], "title": "TestQuestionResult"}, "TestRead": {"properties": {"id": {"type": "string", "title": "Id"}, "questions": {"items": {"$ref": "#/components/schemas/TestQuestionRead"}, "type": "array", "title": "Questions"}}, "type": "object", "required": ["id", "questions"], "title": "TestRead"}, "TherapistCreate": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "age": {"type": "integer", "title": "Age"}, "location": {"type": "string", "title": "Location"}, "education": {"type": "string", "title": "Education"}, "experience": {"type": "integer", "title": "Experience"}, "verified": {"type": "boolean", "title": "Verified", "default": true}, "expertises": {"items": {"type": "integer"}, "type": "array", "title": "Expertises", "default": []}}, "type": "object", "required": ["user_id", "first_name", "last_name", "email", "age", "location", "education", "experience"], "title": "TherapistCreate"}, "TherapistProfileView": {"properties": {"id": {"type": "string", "title": "Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Email"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "age": {"type": "integer", "title": "Age"}, "education": {"type": "string", "title": "Education"}, "experience": {"type": "integer", "title": "Experience"}, "location": {"type": "string", "title": "Location"}, "expertises": {"items": {"type": "integer"}, "type": "array", "title": "Expertises"}}, "type": "object", "required": ["id", "first_name", "last_name", "email", "phone", "age", "education", "experience", "location", "expertises"], "title": "TherapistProfileView"}, "TherapistUserRead": {"properties": {"created_at": {"type": "string", "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}, "id": {"type": "string", "title": "Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "age": {"type": "integer", "title": "Age"}, "location": {"type": "string", "title": "Location"}, "role": {"$ref": "#/components/schemas/UserRole"}, "education": {"type": "string", "title": "Education"}, "experience": {"type": "integer", "title": "Experience"}, "verified": {"type": "boolean", "title": "Verified"}, "expertises": {"items": {"type": "integer"}, "type": "array", "title": "Expertises"}}, "type": "object", "required": ["created_at", "updated_at", "id", "first_name", "last_name", "email", "phone", "age", "location", "role", "education", "experience", "verified", "expertises"], "title": "TherapistUserRead"}, "TherapistUserUpdate": {"properties": {"first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "age": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Age"}, "location": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Location"}, "education": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Education"}, "experience": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Experience"}, "expertises": {"anyOf": [{"items": {"type": "integer"}, "type": "array"}, {"type": "null"}], "title": "Expertises"}}, "type": "object", "title": "TherapistUserUpdate"}, "UserRole": {"type": "string", "enum": ["Military", "Therapist", "Admin"], "title": "UserRole"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}